Changed: Using epoll insted of poll.
Epoll is far more efficient to multiplex the sockets.
Now, it can handle a greater number of users.

Other changes: Fixed a bug when the http connection was closed
before the response was sent, this could cause nasty errors.
Allow \n as a line separator on the http header, this made
testing a lot easier.